moved hiddenservice methods
